When moving a file, we now copy it to the new location, then delete the old one. Doing it this way reduces the complexity; however, this will likely be a bit less efficient.
When a function in file_system.py needs to make a call to swift it now enters an
internal queue that checks to make sure another operation is not already in progress
for that file. If there is another operation in progress we simply put it back on the
queue and wait until we can proceed. This has the advantage of not blocking the main
event thread. (though this can definitely be improved).

Removed functions for moving an object from swift_source and swift_worker